# docs/openapi.yml
{
  "openapi": "3.0.2",
  "info": {
    "title": "Observer API",
    "description": "Observer API server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/pets": {
      "post": {
        "tags": [
          "pets"
        ],
        "summary": "Create Pet",
        "operationId": "create_pet_pets_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pets/{pet_id}": {
      "get": {
        "tags": [
          "pets"
        ],
        "summary": "Get Pet",
        "operationId": "get_pet_pets__pet_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Pet Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "pet_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pets"
        ],
        "summary": "Update Pet",
        "operationId": "update_pet_pets__pet_id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Pet Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "pet_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pets"
        ],
        "summary": "Delete Pet",
        "operationId": "delete_pet_pets__pet_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Pet Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "pet_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pets/{pet_id}/document": {
      "post": {
        "tags": [
          "pets",
          "documents"
        ],
        "summary": "Pet Upload Document",
        "operationId": "pet_upload_document_pets__pet_id__document_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Pet Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "pet_id",
            "in": "path"
          },
          {
            "description": "Uploaded file size",
            "required": true,
            "schema": {
              "title": "Content-Length",
              "type": "integer",
              "description": "Uploaded file size"
            },
            "name": "content-length",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pet_upload_document_pets__pet_id__document_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pets/{pet_id}/documents": {
      "get": {
        "tags": [
          "pets",
          "documents"
        ],
        "summary": "Pet Get Documents",
        "operationId": "pet_get_documents_pets__pet_id__documents_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Pet Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "pet_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Pet Get Documents Pets  Pet Id  Documents Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "title": "APIError",
        "required": [
          "code",
          "status_code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "data": {
            "title": "Data"
          }
        }
      },
...
      "Body_pet_upload_document_pets__pet_id__document_post": {
        "title": "Body_pet_upload_document_pets__pet_id__document_post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "DocumentResponse": {
        "title": "DocumentResponse",
        "required": [
          "name",
          "size",
          "mimetype",
          "owner_id",
          "id",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Document filename"
          },
          "size": {
            "title": "Size",
            "exclusiveMinimum": 0.0,
            "type": "integer",
            "description": "Size of document in bytes"
          },
          "mimetype": {
            "title": "Mimetype",
            "type": "string",
            "description": "Document type (mimetype)"
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Owner ID"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project ID"
          },
          "id": {
            "title": "Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Document ID"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "description": "Creation date and time",
            "format": "date-time"
          }
        }
      },
      "ErrorCode": {
        "title": "ErrorCode",
        "enum": [
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict_error",
          "totp_error",
          "totp_invalid_backup_code_error",
          "totp_required_error",
          "totp_exists_error",
          "registration_error",
          "registrations_closed_error",
          "password_reset_code_expired_error",
          "weak_password_error",
          "invalid_password_error",
          "similar_passwords_error",
          "document_is_too_large_error",
          "unsupported_document_format",
          "confirmation_code_expired_error",
          "invite_expired_error",
          "internal_error",
          "bad_request"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "NewPetRequest": {
        "title": "NewPetRequest",
        "required": [
          "name",
          "status",
          "owner_id",
          "project_id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Pet's name"
          },
          "notes": {
            "title": "Notes",
            "type": "string",
            "description": "Additional notes"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PetStatus"
              }
            ],
            "description": "Pet status"
          },
          "registration_id": {
            "title": "Registration Id",
            "type": "string",
            "description": "Pet's registration ID, passport ID etc."
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Pet's owner ID (Person ID)"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project owner ID"
          }
        }
      },
...
      "PetResponse": {
        "title": "PetResponse",
        "required": [
          "id",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Pet ID"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "description": "Creation datetime",
            "format": "date-time"
          }
        }
      },
      "PetStatus": {
        "title": "PetStatus",
        "enum": [
          "registered",
          "adopted",
          "owner_found",
          "needs_shelter",
          "unknown"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "UpdatePetRequest": {
        "title": "UpdatePetRequest",
        "required": [
          "name",
          "status",
          "owner_id",
          "project_id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Pet's name"
          },
          "notes": {
            "title": "Notes",
            "type": "string",
            "description": "Additional notes"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PetStatus"
              }
            ],
            "description": "Pet status"
          },
          "registration_id": {
            "title": "Registration Id",
            "type": "string",
            "description": "Pet's registration ID, passport ID etc."
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Pet's owner ID (Person ID)"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project owner ID"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}